chore: refactor and add tests for RouterFallback
