+++ src/components/app/Root.jsx
@@ -1,8 +1,7 @@
 import {
-  Outlet, ScrollRestoration, useFetchers, useNavigation, useParams,
+  Outlet, ScrollRestoration, useParams,
 } from 'react-router-dom';
-import { Suspense, useContext, useEffect } from 'react';
-import NProgress from 'accessible-nprogress';
+import { Suspense, useContext } from 'react';
 import { AppContext } from '@edx/frontend-platform/react';
 import { getConfig } from '@edx/frontend-platform';
 import { getLoginRedirectUrl } from '@edx/frontend-platform/auth';
@@ -12,30 +11,7 @@ import DelayedFallbackContainer from '../DelayedFallback/DelayedFallbackContaine
 import { Toasts, ToastsProvider } from '../Toasts';
 import NoticesProvider from '../notices-provider';
 import { ErrorPage } from '../error-page';
-
-// Determines amount of time that must elapse before the
-// NProgress loader is shown in the UI. No need to show it
-// for quick route transitions.
-export const NPROGRESS_DELAY_MS = 300;
-
-export function useNProgressLoader() {
-  const { authenticatedUser } = useContext(AppContext);
-  const navigation = useNavigation();
-  const fetchers = useFetchers();
-
-  useEffect(() => {
-    const timeoutId = setTimeout(() => {
-      const fetchersIdle = fetchers.every((f) => f.state === 'idle');
-      const isAuthenticatedUserHydrated = !!authenticatedUser?.profileImage;
-      if (navigation.state === 'idle' && fetchersIdle && isAuthenticatedUserHydrated) {
-        NProgress.done();
-      } else {
-        NProgress.start();
-      }
-    }, NPROGRESS_DELAY_MS);
-    return () => clearTimeout(timeoutId);
-  }, [navigation, fetchers, authenticatedUser]);
-}
+import { useNProgressLoader } from './data';
 
 const Root = () => {
   const { authenticatedUser } = useContext(AppContext);
