+++ src/components/app/routes/RouterFallback.jsx
@@ -1,0 +1,8 @@
+import { useNProgressLoader } from '../data';
+
+const RouterFallback = () => {
+  useNProgressLoader();
+  return null;
+};
+
+export default RouterFallback;
