+++ src/components/app/routes/RouterFallback.test.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import '@testing-library/jest-dom/extend-expect';
+
+import RouterFallback from './RouterFallback';
+import { useNProgressLoader } from '../data';
+
+jest.mock('../data', () => ({
+  ...jest.requireActual('../data'),
+  useNProgressLoader: jest.fn(),
+}));
+
+describe('RouterFallback', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('renders without crashing', () => {
+    const { container } = render(<RouterFallback />);
+    expect(container).toBeEmptyDOMElement();
+    expect(useNProgressLoader).toHaveBeenCalledTimes(1);
+  });
+});
