+++ src/components/app/routes/index.js
@@ -1,7 +1,9 @@
 export { default as CourseRoute } from './CourseRoute';
 export { default as DashboardRoute } from './DashboardRoute';
 export { default as SearchRoute } from './SearchRoute';
+export { default as RouterFallback } from './RouterFallback';
 export { default as RouteErrorBoundary } from './RouteErrorBoundary';
+export { default as createAppRouter } from './createAppRouter';
 
 export * from './loaders';
 export * from './data';
